(Logical change 1.4102)

+++ plugin/item/stat40/stat40.h
@@ -14,7 +14,7 @@
 
 /* Type for stat40 layout callback function */
 typedef errno_t (*ext_func_t) (sdext_entity_t *,
-			       uint16_t, void *);
+			       uint64_t, void *);
 
 extern errno_t stat40_traverse(reiser4_place_t *place,
 			       ext_func_t ext_func,
